Fixing Windows build issues for a previous commit: 1. model/pattern.h now needs to include <bitset> and <vector>     (because it uses std::bitset and std::vector). 2. In Alignment::constructPatterns the omp parallel for to calculate     total_singleton_parsimony_states needs to use a local variable     (limitations of Microsoft's Open MP implementation). 3.  In Alignment::computeDivergenceMatrix, a static_cast      (to double) was needed to silence a warning about possible      loss of precision.  Won't happen, unless a site has a frequency      of 2^50 or so!  Not likely! 4. In Pattern::countTowardSingletonParsimonyStates, added a static_cast     (to int) for # of states, and changed the type of last_taxon_id to size_t.     (both these changes are to avoid warnings).

+++ alignment/pattern.cpp
@@ -212,11 +212,11 @@ void Pattern::setInformativeFlags(Alignment* aln) {
 }
 
 void Pattern::countTowardSingletonParsimonyStates(std::vector<UINT>& singleton_parsimony_states) const {
-    int num_states = num_app.size();
+    int num_states = static_cast<int>(num_app.size());
     if (!is_informative && 0<count_singleton) {
         for (int j = 0; j < num_states; j++) {
             if (num_app[j]==1) {
-                int last_taxon_id = last_app[j];
+                size_t last_taxon_id = last_app[j];
                 singleton_parsimony_states[last_taxon_id] += frequency;
             }
         }
